修正JWT过期时间过长，VUE客户端登录用户名显示的问题

+++ IoTSharp/AppSettings.cs
@@ -52,7 +52,7 @@ namespace IoTSharp
         public string JwtKey { get; set; }
         public string JwtIssuer { get; set; }
         public string JwtAudience { get; set; }
-        public int JwtExpireHours { get; set; }
+        public double JwtExpireHours { get; set; }
         /// <summary>
         /// Broker settings
         /// </summary>
